+++ include/Renderable/Primitives/SemiCircle.hpp
@@ -11,7 +11,7 @@ namespace Lexvi {
 
     struct SemiCircleMesh {
         std::vector<SemiCircleVertex> vertices;
-        std::vector<unsigned int> indices;
+        std::vector<uint16_t> indices;
         unsigned int VAO = 0, VBO = 0, EBO = 0;
     };
